float32

+++ crates/metarustistic-functions/src/lib.rs
@@ -1,10 +1,10 @@
 
-use std::f64::consts::PI;
+use std::f32::consts::PI;
 
-pub fn rastrigin(vector: &Vec<f64>) -> f64 {
-    let mut r: f64 = 10.0 * vector.len() as f64;
-    vector.iter().for_each(|x: &f64| {
+pub fn rastrigin(vector: &Vec<f32>) -> f32 {
+    let mut r: f32 = 10.0 * vector.len() as f32;
+    vector.iter().for_each(|x: &f32| {
         r += x.powf(2.0) - 10.0 * (2.0 * PI * x).cos();
     });
-    return r as f64;
+    return r as f32;
 }
