+++ src/main.rs
@@ -8,15 +8,15 @@ fn main() {
         step_number:1,
     };
     let start = Instant::now();
-    let mut min:f64=80.0;
-    let mut b_traj:Vec<Vec<f64>>=vec![];
+    let mut min:f32=80.0;
+    let mut b_traj:Vec<Vec<f32>>=vec![];
     for _i in 0..1000000{
         opti.step();
         let r = rastrigin(&opti.positions);
         if r<min{
             min = r;
             println!("Best at step {} {}",opti.step_number,min);
-            let point:Vec<f64> = vec![opti.positions[0],opti.positions[1],r];
+            let point:Vec<f32> = vec![opti.positions[0],opti.positions[1],r];
             b_traj.push(point);
 
         }
